requirements.txt
Django==5.2.6
Pillow==11.3.0
requests==2.32.5
python-decouple==3.8
feedparser==6.0.11
beautifulsoup4==4.13.5
yfinance
